%[dont read]

\name{rangen-package}
\alias{rangen-package}
\docType{package}
\title{
Really fast R functions
}

\description{
A collection of rangen functions for data analysis. 
}

\details{
\tabular{ll}{
Package: \tab \packageName{rangen}\cr
Type: \tab \packageType{rangen}\cr
Version: \tab \packageVersion{rangen} \cr
Date: \tab \packageDate{rangen}\cr
License: \tab \packageLicense{rangen}\cr
}
}

\author{
\itemize{
\item Manos Papadakis <papadakm95@gmail.com>
\item Michail Tsagris <mtsagris@uoc.gr>
}
}

\section{Maintainers }{
Maintainer: \packageMaintainer{rangen}
}

%\note{
%Acknowledgments:
%}

