%[dont read]

\name{rangen-package}
\alias{rangen-package}
\docType{package}
\title{
Random Number Generators and Utilities
}

\description{
Provides a collection of random number generators for common and custom distributions, along with utility functions for sampling and simulation.
}

\details{
\tabular{ll}{
Package: \tab \packageName{rangen}\cr
Type: \tab \packageType{rangen}\cr
Version: \tab \packageVersion{rangen} \cr
Date: \tab \packageDate{rangen}\cr
License: \tab \packageLicense{rangen}\cr
}
}

\author{
\itemize{
\item Manos Papadakis <papadakm95@gmail.com>
\item Michail Tsagris <mtsagris@uoc.gr>
}
}

\section{Maintainers }{
\packageMaintainer{rangen}
}