\name{rangen-package}
\alias{rangen-package}
\alias{rangen}
\docType{package}
\title{
\packageTitle{rangen}
}
\description{
\packageDescription{rangen}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{rangen}

Maintainer: \packageMaintainer{rangen}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = rangen)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{rangen}}
%\section{Documentation Index}{\packageIndices{rangen}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
