\name{ragen-package}
\alias{ragen-package}
\alias{ragen}
\docType{package}
\title{
\packageTitle{ragen}
}
\description{
\packageDescription{ragen}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{ragen}

Maintainer: \packageMaintainer{ragen}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = ragen)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{ragen}}
%\section{Documentation Index}{\packageIndices{ragen}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
